requirements_backup.txt
absl-py==2.3.1
accelerate==1.10.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
basicsr==1.4.2
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
diffusers==0.35.1
easydict==1.13
facexlib==0.3.0
filelock==3.19.1
filterpy==1.4.5
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.74.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
llvmlite==0.44.0
lmdb==1.7.3
lpips==0.1.4
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.2
realesrgan==0.3.0
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tb-nightly==2.21.0a20250911
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.0
tomli==2.2.1
torch==2.8.0
torchaudio==2.8.0
torchmetrics==1.8.2
torchvision==0.12.0
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
